requirements.txt
appdirs==1.4.3
Django==1.11
django-debug-toolbar==1.7
django-dotenv==1.4.1
django-filter==1.0.2
django-model-utils==3.0.0
django-redis-cache==1.7.1
django-secure==1.0.1
django-waffle==0.11.1
djangorestframework==3.6.2
Markdown==2.6.8
packaging==16.8
paho-mqtt==1.2.3
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
requests==2.13.0
six==1.10.0
South==1.0.2
sqlparse==0.2.3
